Add AsyncSequence conformance to PaginatedRows
Motivation:
`PaginatedRows` has `forEach` and `map` APIs for iterating and transforming all rows in the result. Their implementation uses recusion however and probably not the most performant.

Modifications:
Conform `PaginatedRows` to `AsyncSequence` with underlying async stream to take advantage of language constructs and optimization.

Deprecate `EventLoopFuture` and async/await variants of `forEach` and `map`.
