+++ Sources/CassandraClient/Data.swift
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift Cassandra Client open source project
 //
-// Copyright (c) 2022 Apple Inc. and the Swift Cassandra Client project authors
+// Copyright (c) 2022-2023 Apple Inc. and the Swift Cassandra Client project authors
 // Licensed under Apache License v2.0
 //
 // See LICENSE.txt for license information
@@ -102,164 +102,6 @@ extension CassandraClient {
                 return Row(iterator: self)
             }
         }
-    }
-
-    /// Resulting row(s) of a Cassandra query. Data are paginated.
-    public final class PaginatedRows {
-        let session: Session
-        let statement: Statement
-        let eventLoop: EventLoop?
-        let logger: Logger?
-
-        /// If `true`, calling ``nextPage()-4komz`` will return a new set of ``CassandraClient/Rows``.
-        /// Otherwise it will throw ``CassandraClient/Error/rowsExhausted`` error.
-        public private(set) var hasMorePages: Bool = true
-
-        internal init(session: Session, statement: Statement, on eventLoop: EventLoop, logger: Logger?) {
-            self.session = session
-            self.statement = statement
-            self.eventLoop = eventLoop
-            self.logger = logger
-        }
-
-        /// Fetches next page of rows or throw ``CassandraClient/Error/rowsExhausted`` error if there are no more pages.
-        public func nextPage() -> EventLoopFuture<Rows> {
-            guard let eventLoop = self.eventLoop else {
-                preconditionFailure("EventLoop must not be nil")
-            }
-            guard self.hasMorePages else {
-                return eventLoop.makeFailedFuture(CassandraClient.Error.rowsExhausted)
-            }
-
-            let future = self.session.execute(statement: self.statement, on: eventLoop, logger: self.logger)
-            future.whenComplete { result in
-                switch result {
-                case .success(let rows):
-                    self.hasMorePages = cass_result_has_more_pages(rows.rawPointer) == cass_true
-                    if self.hasMorePages {
-                        cass_statement_set_paging_state(self.statement.rawPointer, rows.rawPointer)
-                    }
-                case .failure:
-                    self.hasMorePages = false
-                }
-            }
-            return future
-        }
-
-        /// Iterates through all rows in all pages and invokes the given closure on each.
-        public func forEach(_ body: @escaping (Row) throws -> Void) -> EventLoopFuture<Void> {
-            precondition(self.hasMorePages, "Only one of 'forEach' or 'map' can be called once per PaginatedRows")
-
-            guard let eventLoop = self.eventLoop else {
-                preconditionFailure("EventLoop must not be nil")
-            }
-
-            func _forEach() -> EventLoopFuture<Void> {
-                self.nextPage().flatMap { rows in
-                    do {
-                        try rows.forEach(body)
-                    } catch {
-                        return eventLoop.makeFailedFuture(error)
-                    }
-
-                    guard self.hasMorePages else {
-                        return eventLoop.makeSucceededFuture(())
-                    }
-                    return _forEach()
-                }
-            }
-            return _forEach()
-        }
-
-        /// Iterates through all rows in all pages and applies `transform` on each.
-        public func map<T>(_ transform: @escaping (Row) throws -> T) -> EventLoopFuture<[T]> {
-            precondition(self.hasMorePages, "Only one of 'forEach' or 'map' can be called once per PaginatedRows")
-
-            guard let eventLoop = self.eventLoop else {
-                preconditionFailure("EventLoop must not be nil in EventLoop based APIs")
-            }
-
-            func _map(_ accumulated: [T]) -> EventLoopFuture<[T]> {
-                self.nextPage().flatMap { rows in
-                    do {
-                        let transformed = try rows.map(transform)
-                        let newAccumulated = accumulated + transformed
-                        guard self.hasMorePages else {
-                            return eventLoop.makeSucceededFuture(newAccumulated)
-                        }
-                        return _map(newAccumulated)
-                    } catch {
-                        return eventLoop.makeFailedFuture(error)
-                    }
-                }
-            }
-            return _map([])
-        }
-
-        #if compiler(>=5.5) && canImport(_Concurrency)
-        @available(macOS 12, iOS 15, tvOS 15, watchOS 8, *)
-        internal init(session: Session, statement: Statement, logger: Logger?) {
-            self.session = session
-            self.statement = statement
-            self.eventLoop = nil
-            self.logger = logger
-        }
-
-        /// Fetches next page of rows or throw ``CassandraClient/Error/rowsExhausted`` error if there are no more pages.
-        @available(macOS 12, iOS 15, tvOS 15, watchOS 8, *)
-        public func nextPage() async throws -> Rows {
-            guard self.hasMorePages else {
-                throw CassandraClient.Error.rowsExhausted
-            }
-
-            do {
-                let rows = try await session.execute(statement: self.statement, logger: self.logger)
-                self.hasMorePages = cass_result_has_more_pages(rows.rawPointer) == cass_true
-                if self.hasMorePages {
-                    cass_statement_set_paging_state(self.statement.rawPointer, rows.rawPointer)
-                }
-                return rows
-            } catch {
-                self.hasMorePages = false
-                throw error
-            }
-        }
-
-        /// Iterates through all rows in all pages and invokes the given closure on each.
-        @available(macOS 12, iOS 15, tvOS 15, watchOS 8, *)
-        public func forEach(_ body: @escaping (Row) throws -> Void) async throws {
-            precondition(self.hasMorePages, "Only one of 'forEach' or 'map' can be called once per PaginatedRows")
-
-            func _forEach() async throws {
-                let rows = try await nextPage()
-                try rows.forEach(body)
-
-                guard self.hasMorePages else {
-                    return
-                }
-                try await _forEach()
-            }
-            try await _forEach()
-        }
-
-        /// Iterates through all rows in all pages and applies `transform` on each.
-        @available(macOS 12, iOS 15, tvOS 15, watchOS 8, *)
-        public func map<T>(_ transform: @escaping (Row) throws -> T) async throws -> [T] {
-            precondition(self.hasMorePages, "Only one of 'forEach' or 'map' can be called once per PaginatedRows")
-
-            func _map(_ accumulated: [T]) async throws -> [T] {
-                let rows = try await nextPage()
-                let transformed = try rows.map(transform)
-                let newAccumulated = accumulated + transformed
-
-                guard self.hasMorePages else {
-                    return newAccumulated
-                }
-                return try await _map(newAccumulated)
-            }
-            return try await _map([])
-        }
-        #endif
     }
 
     /// A resulting row of a Cassandra query.
